app/requirements.txt
aioredis==2.0.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
pwdlib==0.2.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.1.1
pytest-ordering==0.6
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
